♻️ [refactor] tour 페이지 섹션 별 분기
